removed plot function from aquire utils as it lead to circular imports
