+++ tests/core_tests/measurement_tests/aquire_test.py
@@ -545,9 +545,3 @@ class TestAquire:
         model.info(name="circuits")
         fig, ax = model.plot()
 
-
-
-
-
-
-
